requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
itsdangerous==2.2.0
python-multipart==0.0.20
sqlalchemy==2.0.43
pymysql==1.1.1
alembic==1.16.4
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator==2.2.0
python-dotenv==1.1.1
gunicorn==23.0.0
httpx==0.28.1
pytest==8.4.1
